replace old geometry with new one
