--- src/modules/dream/test/DreamGeometryIITest.cpp
@@ -1,88 +1,0 @@
-// Copyright (C) 2021 European Spallation Source, see LICENSE file
-//===----------------------------------------------------------------------===//
-///
-/// \file
-///
-/// \brief Unit test for Dream geometry
-///
-//===----------------------------------------------------------------------===//
-
-#include <test/TestBase.h>
-#include <dream/geometry/Geometry.h>
-
-using namespace Dream;
-
-class DreamGeometryTest : public TestBase {
-protected:
-  const uint8_t Sector0{0};
-  const uint8_t Sumo0{0}; // invalid. Valid range is 3 - 6
-  const uint8_t Sumo3{3};
-  const uint8_t Cassette0{0};
-  const uint8_t Counter0{0};
-  const uint8_t Wire0{0};
-  const uint8_t Strip0{0};
-  EndCapGeometry endcap;
-  void SetUp() override {}
-  void TearDown() override {}
-};
-
-TEST_F(DreamGeometryTest, Constructor) {
-  ASSERT_EQ(endcap.getPixel(Sector0, Sumo0, Cassette0, Counter0, Wire0, Strip0), 0);
-}
-
-TEST_F(DreamGeometryTest, InvalidSector) {
-  ASSERT_EQ(endcap.getPixel(23, Sumo3, Cassette0, Counter0, Wire0, Strip0), 0);
-  ASSERT_EQ(endcap.getPixel(23, Sumo3, Cassette0, Counter0, Wire0, Strip0), 0);
-  ASSERT_EQ(endcap.getPixel(24, Sumo3, Cassette0, Counter0, Wire0, Strip0), 0);
-}
-
-TEST_F(DreamGeometryTest, ValidSector) {
-  for (uint8_t Sector = 0; Sector < 23; Sector++) {
-    ASSERT_NE(endcap.getPixel(Sector, Sumo3, Cassette0, Counter0, Wire0, Strip0), 0);
-  }
-}
-
-TEST_F(DreamGeometryTest, InvalidSumo) {
-  ASSERT_EQ(endcap.getPixel(Sector0, 0, Cassette0, Counter0, Wire0, Strip0), 0);
-  ASSERT_EQ(endcap.getPixel(Sector0, 1, Cassette0, Counter0, Wire0, Strip0), 0);
-  ASSERT_EQ(endcap.getPixel(Sector0, 2, Cassette0, Counter0, Wire0, Strip0), 0);
-  ASSERT_EQ(endcap.getPixel(Sector0, 7, Cassette0, Counter0, Wire0, Strip0), 0);
-  ASSERT_EQ(endcap.getPixel(Sector0, 8, Cassette0, Counter0, Wire0, Strip0), 0);
-}
-
-TEST_F(DreamGeometryTest, ValidSumo) {
-  ASSERT_NE(endcap.getPixel(Sector0, 3, Cassette0, Counter0, Wire0, Strip0), 0);
-  ASSERT_NE(endcap.getPixel(Sector0, 4, Cassette0, Counter0, Wire0, Strip0), 0);
-  ASSERT_NE(endcap.getPixel(Sector0, 5, Cassette0, Counter0, Wire0, Strip0), 0);
-  ASSERT_NE(endcap.getPixel(Sector0, 6, Cassette0, Counter0, Wire0, Strip0), 0);
-}
-
-
-TEST_F(DreamGeometryTest, TestingICD4Corners) {
-  //                        sec su cas ctr wir str
-  ASSERT_EQ(endcap.getPixel( 0, 6,  0,  0,  0,  0),    1); // upper left (z = 0)
-  ASSERT_EQ(endcap.getPixel( 1, 6,  0,  0,  0,  0),   57); // next sector 'upper left'
-  ASSERT_EQ(endcap.getPixel(22, 3,  3,  1,  0,  0), 1288); // upper right=
-  ASSERT_EQ(endcap.getPixel( 0, 6,  0,  0, 15, 15), 328'441); // bottom left
-  ASSERT_EQ(endcap.getPixel(22, 3,  3,  1, 15, 15), 329'728); // bottom right
-}
-
-TEST_F(DreamGeometryTest, TestingICD4BoxCorners) {
-  // Refer to figure 7 (logical geometry)
-  //                    sec su cas ctr
-  ASSERT_EQ(endcap.getX( 0,  6, 0,  0),        0); // top left
-  ASSERT_EQ(endcap.getX(22,  3, 3,  1),     1287); // top right
-  ASSERT_EQ(endcap.getX( 3,  6, 0,  0), 168 +  0); // box left
-  ASSERT_EQ(endcap.getX( 3,  3, 3,  1), 168 + 55); // box right
-
-  //                    wir str
-  ASSERT_EQ(endcap.getY( 0,  3), 48 +  0); // box top
-  ASSERT_EQ(endcap.getY(15,  3), 48 + 15); // box bottom
-  ASSERT_EQ(endcap.getY( 0,  0),       0); // top
-  ASSERT_EQ(endcap.getY(15, 15),     255); // bottom
-}
-
-int main(int argc, char **argv) {
-  testing::InitGoogleTest(&argc, argv);
-  return RUN_ALL_TESTS();
-}
